4010: Removed unnecessary useTourSteps hook
